Refactors auth and user data handling
Refactors authentication data source to repository pattern.
Changes UserStats to use Long for xp and currency.
Adds duration to task DTO and domain models.
Adds level calculation based on XP.
Improves achievement handling to show if achieved.
Updates user repository for item purchasing.
Adds progression service for level calculation

+++ app/src/main/java/marcel/uni/gamifiedplanner/domain/user/model/UserStats.kt
@@ -3,6 +3,6 @@ package marcel.uni.gamifiedplanner.domain.user.model
 data class UserStats(
     val uid: String = "",
     val username: String = "",
-    val xp: Int = 0,
-    val currency: Int = 0,
+    val xp: Long = 0,
+    val currency: Long = 0,
 )
